+++ openilc/__init__.py
@@ -1,11 +1,11 @@
 from .hilc import HILC
 from .nilc import NILC
-from .configs import ConfigTable, load_table, load_yaml_config
+from .configs import ConfigTable, load_csv_table, load_table
 
 __all__ = [
     "ConfigTable",
     "HILC",
     "NILC",
+    "load_csv_table",
     "load_table",
-    "load_yaml_config",
 ]