+++ tutorials/tutorial_ilc_bias.py
@@ -54,7 +54,7 @@ def run_nilc_w_map():
     # do nilc and save the weights in map
     time0 = time.time()
     sim = np.load("./test_data/sim_cn.npy")
-    obj_nilc = NILC.from_config("configs/default.yaml",
+    obj_nilc = NILC.from_csv("configs/bands.csv", "configs/needlets_default.csv",
         weights_name="./nilc_weight/w_map.npz",
         Sm_maps=sim,
         mask=None,
@@ -72,7 +72,7 @@ def run_nilc_w_map():
 def get_n_res_w_map():
     # calc noise bias
     noise = np.load("./test_data/sim_n.npy")
-    obj_nilc = NILC.from_config("configs/default.yaml",
+    obj_nilc = NILC.from_csv("configs/bands.csv", "configs/needlets_default.csv",
         weights_config="./nilc_weight/w_map.npz",
         Sm_maps=noise,
         mask=None,
@@ -88,7 +88,7 @@ def get_n_res_w_map():
 def get_cmb_res_w_map():
     # calc noise bias
     cmb = np.asarray([np.load("./test_data/sim_c.npy")] * 5)
-    obj_nilc = NILC.from_config("configs/default.yaml",
+    obj_nilc = NILC.from_csv("configs/bands.csv", "configs/needlets_default.csv",
         weights_config="./nilc_weight/w_map.npz",
         Sm_maps=cmb,
         mask=None,